feat(services): add BigBlueButton conferencing
Added BigBlueButton as a new service in `services.json`, providing users
with an open-source alternative for video conferencing, webinars, and
more. This addition includes its basic information, operational status,
and a direct link for ease of access. This update aims to enrich the
suite of services with more diverse options, catering to the growing
needs for reliable virtual meeting solutions.