feat(README): add support badge
Introduced a new "Support Us!" badge to the README.md to encourage contributions and support for Private.coffee. This badge not only visually enhances the repository's appeal but also provides a direct link for potential supporters to contribute, fostering a stronger community engagement.

The badge leverages custom styling and is powered by Private.coffee's own badge service, ensuring consistency with the project's branding and identity. This addition aligns with efforts to increase visibility and support for our open-source work.