fix: update membership button text for inclusivity
Changed the text on the membership button from "JOIN & REBEL" to "JOIN & SUPPORT" in the base template. This adjustment better reflects the community spirit and aims to be more inclusive, encouraging broader participation. The update is part of ongoing efforts to make our digital spaces more welcoming and representative of our values.