feat(icons): add LGBTIQ+ rainbow icon
Introduced a new LGBTIQ+ rainbow icon SVG to enhance visual representation and inclusivity on the platform. This update includes the integration of the icon within the base template, specifically at the footer section to signal support for the LGBTIQ+ community. This change aligns with our commitment to diversity and inclusiveness.