feat: add Travelynx service with train icon
Added the new Travelynx service to the services.json file to allow users to log their rail journeys. Introduced a new train icon in SVG format to visually represent this service. This addition enhances the diversity of services offered, catering specifically to users interested in tracking rail journeys in Germany and Europe